将部分头文件作为系统头文件引用

+++ src/system/Linux_Webots/thread.cpp
@@ -1,4 +1,4 @@
-#include "thread.hpp"
+#include <thread.hpp>
 
 #include "bsp_time.h"
 #include "webots/robot.h"
@@ -14,7 +14,7 @@ void Thread::StartKernel() {
 }
 
 void Thread::Sleep(uint32_t microseconds) {
-  float time = bsp_time_get() + microseconds / 1000.0f;
+  float time = bsp_time_get() + static_cast<float>(microseconds) / 1000.0f;
 
   while (bsp_time_get() < time) {
     poll(NULL, 0, 1);
@@ -22,7 +22,7 @@ void Thread::Sleep(uint32_t microseconds) {
 }
 
 void Thread::SleepUntil(uint32_t microseconds) {
-  float time = bsp_time_get() + microseconds / 1000.0f;
+  float time = bsp_time_get() + static_cast<float>(microseconds) / 1000.0f;
 
   while (bsp_time_get() < time) {
     poll(NULL, 0, 1);
